Fix delete category, category cant be deleted if has subcategory

+++ backend/app/Http/Controllers/CategoryController.php
@@ -100,6 +100,14 @@ class CategoryController extends Controller
                 'error' => 'This category is currently in use.'
             ], 422);
         }
+
+        if ($category->children()->exists()) {
+            return response()->json([
+                'message' => 'Cannot delete category because it has subcategory.',
+                'error' => 'This category is currently in use.'
+            ], 422);
+        }
+
         $category->delete();
         return response()->noContent();
     }
